crm-backend/crmapp/requirements.txt
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.0
djangorestframework-simplejwt
pymongo==4.12
werkzeug
